added min and max distance to renderer culling radius. cleaned up code. added tbb init to concurrency. fixed a bug where radius was a uint in shader

+++ concurrency.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <tbb/tbb.h>
 namespace concurrency {
-	extern size_t numThreads;
+	extern int numThreads;
 }
